fix: AutoFocus to focusMe AND Email Regex Change
Updated the first input on every page to have the focusMe directive.

Also, very strange error with the email regex, where the test resut was always console.logged as true, but it wouldn't trip the if test.  When I stored the test bool in a variable and evaluated the variable it worked fine.  Very strange, but it's working now.